changing ikrand functions to take doubles instead of long doubles (even in real*10 version) --Molly
git-svn-id: https://giordano.mit.edu/nvs/mangle/trunk/dev_80bit@290 9f78a35e-8a50-488b-8db1-b351ce30c53d

+++ src/manglefn.h
@@ -184,9 +184,9 @@ int	harmonize_polys(int npoly, polygon *poly[/*npoly*/], long double, int lmax, 
 
 void	harmonizepolys_(long double *, int *, harmonic []);
 
-void	ikrand_(int *, long double *);
-void	ikrandp_(long double *, long double *);
-void	ikrandm_(long double *, long double *);
+void	ikrand_(int *, double *);
+void	ikrandp_(double *, double *);
+void	ikrandm_(double *, double *);
 
 void	msg(char *, ...);
